requirements.txt
django>=3.0, <4.0
djangorestframework>=3.5
django-rest-swagger>=2.2.0
djoser==2.1.0
flake8>=3.9.0
psycopg2-binary>=2.8
python-decouple==3.4
xlrd==1.2.0
